// front/src/shared/api/generated/model/localCartItem.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * FastAPI
 * OpenAPI spec version: 0.1.0
 */
import type { LocalCartItemPriceSnapshot } from './localCartItemPriceSnapshot';

export interface LocalCartItem {
  productId: string;
  quantity: number;
  priceSnapshot: LocalCartItemPriceSnapshot;
}